tests/wpt/tests/css/css-animations/parsing/animation-name-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: getComputedStyle().animationName</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-name">
<meta name="assert" content="animation-name computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("animation-name", 'none');
test_computed_value("animation-name", 'NONE', 'none');

test_computed_value("animation-name", 'foo');
test_computed_value("animation-name", 'Both');
test_computed_value("animation-name", 'ease-in');
test_computed_value("animation-name", 'infinite');
test_computed_value("animation-name", 'paused');
test_computed_value("animation-name", 'first, second, third');

test_computed_value("animation-name", '"something"', 'something');
test_computed_value("animation-name", '"---\\22---"', '---\\\"---');
test_computed_value("animation-name", '"multi word string"', 'multi\\ word\\ string');

// Restricted keywords as strings
test_computed_value("animation-name", '"none"');
test_computed_value("animation-name", '"none", "INITIAL", "inherit"');
test_computed_value("animation-name", '"none", both, ease-in');
test_computed_value("animation-name", '"NoNe"');
test_computed_value("animation-name", '"initial"');
test_computed_value("animation-name", '"INITIAL"');
test_computed_value("animation-name", '"inherit"');
test_computed_value("animation-name", '"INHERIT"');
test_computed_value("animation-name", '"revert"');
test_computed_value("animation-name", '"REVERT"');
test_computed_value("animation-name", '"revert-layer"');
test_computed_value("animation-name", '"REVERT-LAYER"');
test_computed_value("animation-name", '"unset"');
test_computed_value("animation-name", '"UNSET"');
test_computed_value("animation-name", '"default"');
test_computed_value("animation-name", '"DEFAULT"');
</script>
</body>
</html>
